+++ NEWS.rst
@@ -41,6 +41,8 @@ Changes
 * Vendored version of zstd upgraded to 1.3.1.
 * Added module constants ``CONTENTSIZE_UNKNOWN`` and ``CONTENTSIZE_ERROR``.
 * Add ``STRATEGY_BTULTRA`` compression strategy constant.
+* Switch from deprecated ``ZSTD_getDecompressedSize()`` to
+  ``ZSTD_getFrameContentSize()`` replacement.
 
 0.8.1 (released 2017-04-08)
 ---------------------------